Move API management classes to apimanagers package
Having all the classes that managed, or will manage, the PE APIs
in the model's package didn't make any sense. Having a apimanagers
package that all the API management classes go in makes their intent
much clearer.
