+++ src/main/java/org/jenkinsci/plugins/puppetenterprise/apimanagers/PuppetOrchestratorV1/PuppetCommandDeployV1.java
@@ -1,4 +1,4 @@
-package org.jenkinsci.plugins.puppetenterprise.models.puppetorchestratorv1;
+package org.jenkinsci.plugins.puppetenterprise.apimanagers.puppetorchestratorv1;
 
 import java.util.*;
 import java.net.URI;
@@ -8,10 +8,10 @@ import com.google.gson.internal.LinkedTreeMap;
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
 import org.jenkinsci.plugins.workflow.PEException;
-import org.jenkinsci.plugins.puppetenterprise.models.PERequest;
+import org.jenkinsci.plugins.puppetenterprise.apimanagers.PERequest;
 import org.jenkinsci.plugins.puppetenterprise.models.PEResponse;
-import org.jenkinsci.plugins.puppetenterprise.models.PuppetOrchestratorV1;
-import org.jenkinsci.plugins.puppetenterprise.models.puppetorchestratorv1.PuppetOrchestratorException;
+import org.jenkinsci.plugins.puppetenterprise.apimanagers.PuppetOrchestratorV1;
+import org.jenkinsci.plugins.puppetenterprise.apimanagers.puppetorchestratorv1.PuppetOrchestratorException;
 
 public class PuppetCommandDeployV1 extends PuppetOrchestratorV1 {
   private URI uri = null;
